+++ changelog.md
@@ -6,7 +6,10 @@
 
 **FEATURES**
 - Keymap Override: Transform with Gizmos.
-- Operator: Transform Toggle now goes back to the previous tool instead of defaulting to Box Select.
+- New Tool: *Set Custom Orientation* creates a temporaty orientation based on your selection, running the script
+  from an empty area of the viewport goes back to Global pivot orientation.
+- Transform Toggle Tool now goes back to the previous tool instead of defaulting to Box Select.
+- Transform Toggle Tool (now Toggle Tool) now toggles any tool with the input *idname*.
 
 **DEV**
 - Swapped out the keymap functions for classes, simplifying the module's register/unregister methods among other things.